[Add] add logger in TestShell and delegate printing
Green 단계
Unit Test 통과 확인
